+++ src/test/commentTestSamples/issue-85-continue-with-label.go
@@ -1,0 +1,15 @@
+/*
+reach: [["inner", "outer"]]
+*/
+func ContinueToLabel() {
+outer:
+	for i := 0; i < 5; i++ {
+	    // CFG: outer
+		fmt.Println(i)
+		for {
+		    // CFG: inner
+			continue outer
+		}
+		fmt.Println("Oh no!")
+	}
+}