+++ src/test/commentTestSamples/labledStatements.go
@@ -1,9 +1,10 @@
 
 /*
-unreach: [["SOURCE","UNREACH"]]
+unreach: [["SOURCE","UNREACH"], ["SOURCE", "BEFORE_INNER_LOOP"]]
 */
 func LabelledStatementFlow() {
   loop1: for x := range xs {
+    // CFG: BEFORE_INNER_LOOP
     for ;; {
       // CFG: SOURCE
       break loop1
@@ -11,4 +12,20 @@ func LabelledStatementFlow() {
     // CFG: UNREACH
     unreachable_from_inner_loop()
   }
+}
+
+/*
+reach: [["inner", "outer"]]
+*/
+func ContinueToLabel() {
+outer:
+	for i := 0; i < 5; i++ {
+	    // CFG: outer
+		fmt.Println(i)
+		for {
+		    // CFG: inner
+			continue outer
+		}
+		fmt.Println("Oh no!")
+	}
 }