+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - `finally` blocks are now supported in TypeScript. 
 - Methods are now supported in TypeScript.
+- `continue label` is now supported in Go.
 - Exporting DOT files (in debug mode) no longer results in different-looking graphs
 
 ### Changed